fix: Migrating deployment from GCP trial to GitHub Pages

+++ Makefile
@@ -19,21 +19,16 @@ run_app:
 	ls -a pages_files
 	ls -a pages_files/assets
 
-	find pages_files -exec sed -i.bak 's|_dash-component-suites|dash-actions-tutorial\\/_dash-component-suites|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-layout|dash-actions-tutorial/_dash-layout.json|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-dependencies|dash-actions-tutorial/_dash-dependencies.json|g' {} \;
-	find pages_files -exec sed -i.bak 's|_reload-hash|dash-actions-tutorial/_reload-hash|g' {} \;
-	find pages_files -exec sed -i.bak 's|_dash-update-component|dash-actions-tutorial/_dash-update-component|g' {} \;
-	find pages_files -exec sed -i.bak 's|assets|dash-actions-tutorial/assets|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-component-suites|dash-actions-tutorial\\/_dash-component-suites|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-layout|dash-actions-tutorial/_dash-layout.json|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-dependencies|dash-actions-tutorial/_dash-dependencies.json|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_reload-hash|dash-actions-tutorial/_reload-hash|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|_dash-update-component|dash-actions-tutorial/_dash-update-component|g' {} \;
+	find pages_files -type f -exec sed -i.bak 's|assets|dash-actions-tutorial/assets|g' {} \;
 
-	mv pages_files/_dash-layout pages_files/_dash-layout.json
-	mv pages_files/_dash-dependencies pages_files/_dash-dependencies.json
-	mv assets/* pages_files/assets/
+	[ -f pages_files/_dash-layout ] && mv pages_files/_dash-layout pages_files/_dash-layout.json
+	[ -f pages_files/_dash-dependencies ] && mv pages_files/_dash-dependencies pages_files/_dash-dependencies.json
+
+	rsync -av assets/ pages_files/assets/
 
 	ps -C python -o pid= | xargs kill -9
-
-clean_dirs:
-	ls
-	rm -rf 127.0.0.1:8050/
-	rm -rf pages_files/
-	rm -rf joblib
